destructured props, moved styles to styles.js
